+++ 7_distributed_systems/kvstore/data/kvstore.json
@@ -1,1 +1,1 @@
-{"foo":"bar"}
+{"foo":"bar","run":"lift"}
